move : 파일 분리 및 정리
작업 완료
코드가 너무길어 파일로 분리

작업 예정
기본정보를 보여주는 UI 갖추기
특성, 각인 등
UI 전체 정리
